ISSUE-9: extract expand_env_var into separate module
