+++ src/environment/mod.rs
@@ -1,7 +1,7 @@
 use std::env;
 use std::path::{PathBuf, Path};
 
-use regex::{Captures, Regex};
+pub mod expand_env;
 
 pub struct Environment {
     executable_name: String,
@@ -73,22 +73,6 @@ pub fn system_environment() -> Environment {
     };
 }
 
-pub fn expand_env_var(path: &str) -> String {
-    let re = Regex::new(r"(\$\{[^{}]+\})").unwrap();
-    let expanded = re.replace_all(
-        path,
-        |captures: &Captures| -> String {
-            let env_var = captures
-                .get(1)
-                .unwrap()
-                .as_str()
-                .to_string();
-            return env::var(&env_var[2..(env_var.len() - 1)])
-                .unwrap_or(env_var);
-        });
-    return expanded.into_owned();
-}
-
 pub fn autodetect_executable(executable_path: &Path,
                              executable_name: &str,
                              checker: &dyn CheckFile) -> Option<String> {
@@ -146,19 +130,6 @@ mod tests {
 
     use super::*;
     use std::path::Path;
-
-    #[test]
-    fn expand_existing_var() {
-        env::set_var("ENV_VAR", "yes");
-        assert_eq!("yes/replaced",
-                   expand_env_var("${ENV_VAR}/replaced"));
-    }
-
-    #[test]
-    fn not_existing_var_wait_unmodified_string() {
-        assert_eq!("${NOT_EXISTING_VAR}/not_replaced",
-                   expand_env_var("${NOT_EXISTING_VAR}/not_replaced"));
-    }
 
     #[test]
     fn target_executable_can_be_found_later_in_the_path() {
