+++ src/handler/default.rs
@@ -1,7 +1,7 @@
 use crate::{config, environment, process};
 use crate::config::Alias::{RegularAlias, ShellAlias};
 use crate::config::Configuration;
-use crate::environment::{Environment, expand_env_var, autodetect_executable, OsCheckFile};
+use crate::environment::{Environment, expand_env, autodetect_executable, OsCheckFile};
 use crate::handler::Handler;
 use crate::process::CallContext;
 
@@ -9,7 +9,7 @@ fn get_call_context(environment: &environment::Environment,
                     configuration: &config::Configuration) -> Result<CallContext, String> {
     let call_arguments = environment.call_arguments();
     let executable = configuration.get_executable()?
-        .map(|config| expand_env_var(&config))
+        .map(|config| expand_env::expand_env_var(&config))
         .or_else(|| autodetect_executable(
             environment.executable_dir().as_path(),
             environment.executable_name().as_str(),
